backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0
pydantic==2.7.1
python-dotenv==1.0.1

# Document processing
PyPDF2==3.0.1
langchain-text-splitters==0.2.1

# Embeddings & vector store
sentence-transformers==3.0.0
faiss-cpu==1.8.0
numpy==1.26.4
